--- app/components/PostList/styles.ts
@@ -1,10 +1,0 @@
-export const Container = {
-  display: "grid",
-  gap: "1.5rem",
-};
-
-export const CardWrapper = {
-  textDecoration: "none",
-  color: "initial",
-  width: "fit-content",
-};
