+++ app/styles/main.ts
@@ -1,5 +1,5 @@
 import { color } from "@styles/color";
-import { size, mq } from "@styles/size";
+import { SIZE, mq } from "@styles/size";
 
 export const background = {
   display: "flex",
@@ -29,12 +29,12 @@ export const recentPostsConatiner = {
   overflow: "auto",
 
   [mq[0]]: {
-    paddingTop: `calc(${size.TOPBAR_HEIGHT} + 1rem)`,
+    paddingTop: `calc(${SIZE.TOPBAR_HEIGHT} + 1rem)`,
   },
 
   [mq[1]]: {
     padding: "1rem",
-    paddingTop: `calc(${size.TOPBAR_HEIGHT} + 1rem)`,
+    paddingTop: `calc(${SIZE.TOPBAR_HEIGHT} + 1rem)`,
   },
 };
 export const title = {
