+++ app/styles/size.ts
@@ -1,6 +1,14 @@
-export const size = {
+export const SIZE = {
   TOPBAR_HEIGHT: "3rem",
   CATEGORY_WIDTH: "18rem",
+
+  BORDER: {
+    POST_CARD: "2px",
+  },
+
+  BORDER_RADIUS: {
+    POST_CARD: "1rem",
+  },
 };
 
 const BREAK_POINTS = [1200, 576];
